Release 1.1.0

+++ src/main/java/edu/eci/cvds/pattens/service/TaskService.java
@@ -7,8 +7,9 @@ import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.TransactionSystemException;
 import org.springframework.transaction.annotation.Transactional;
-
+import java.time.LocalDate;
 import java.util.List;
+import java.util.Random;
 
 /**
  * Service class for managing tasks.
@@ -52,17 +53,23 @@ public class TaskService {
      * @throws Exception if an error occurs during creation
      */
     @Transactional
-    public Task createTask(Task task) throws Exception {
+    public Task createTask(Task task) throws Exception, DataIntegrityViolationException, RuntimeException {
         try {
-            //if the id lready exists, throw an exception
-            if (taskRepository.existsById(String.valueOf(task.getId()))) {
+            if (taskRepository.existsById(task.getId())) {
                 throw new DataIntegrityViolationException("Task already exists");
+            }
+            if (!(task.getDifficultyLevel().equalsIgnoreCase("low") || task.getDifficultyLevel().equalsIgnoreCase("medium" )|| task.getDifficultyLevel().equalsIgnoreCase("high"))) {
+                throw new DataIntegrityViolationException("Invalid difficulty level");
+            }
+            if (task.getPriority() < 1 || task.getPriority() > 5) {
+                throw new DataIntegrityViolationException("Invalid priority");
+            }
+            if(task.getEstimatedTime().isBefore(LocalDate.now())){
+                throw new RuntimeException("Invalid time");
             }
             return taskRepository.saveTask(task);
         } catch (TransactionSystemException e) {
             throw new TransactionSystemException("Error creating task");
-        } catch (Exception e) {
-            throw new Exception("Error creating task");
         }
     }
 
@@ -76,15 +83,21 @@ public class TaskService {
     @Transactional
     public Task updateTask(Task task) throws Exception {
         try {
-            if (taskRepository.existsById(String.valueOf(task.getId()))) {
-                return taskRepository.updateTask(task);
-            } else {
-                throw new Exception("Task not found");
+            if (!taskRepository.existsById(task.getId())) {
+                throw new DataIntegrityViolationException("Task not found");
             }
+            if (!(task.getDifficultyLevel().equalsIgnoreCase("low") || task.getDifficultyLevel().equalsIgnoreCase("medium" )|| task.getDifficultyLevel().equalsIgnoreCase("high"))) {
+                throw new DataIntegrityViolationException("Invalid difficulty level");
+            }
+            if (task.getPriority() < 1 || task.getPriority() > 5) {
+                throw new DataIntegrityViolationException("Invalid priority");
+            }
+            if(task.getEstimatedTime().isBefore(LocalDate.now())){
+                throw new RuntimeException("Invalid time");
+            }
+            return taskRepository.updateTask(task);
         } catch (TransactionSystemException e) {
-            throw new TransactionSystemException("Transaction system error", e);
-        } catch (Exception e) {
-            throw new Exception("An unexpected error occurred", e);
+            throw new TransactionSystemException("Error creating task");
         }
     }
 
@@ -95,6 +108,9 @@ public class TaskService {
      */
     public void deleteTask(String id) {
         Task task = getTaskById(id);
+        if(task == null){
+            throw new RuntimeException("Task not found");
+        }
         taskRepository.deleteTask(task);
     }
 
@@ -105,6 +121,9 @@ public class TaskService {
      */
     public Task doneTask(String id) {
         Task task = getTaskById(id);
+        if(task == null){
+            throw new RuntimeException("Task not found");
+        }
         if(task.getIsCompleted()){
             return task;
         }
@@ -119,6 +138,9 @@ public class TaskService {
      */
     public Task undoneTask(String id) {
         Task task = getTaskById(id);
+        if(task == null){
+            throw new RuntimeException("Task not found");
+        }
         if (!task.getIsCompleted()) {
             return task;
         }
@@ -135,7 +157,77 @@ public class TaskService {
      */
     public Task changeIsCompleted(String id) {
         Task task = getTaskById(id);
+        if(task == null){
+            throw new RuntimeException("Task not found");
+        }
         task.setIsCompleted(!task.getIsCompleted());
+        if (task.getFinishDate() == null) {
+            task.setFinishDate(LocalDate.now());
+        } else {
+            task.setFinishDate(null);
+        }
         return taskRepository.updateTask(task);
+    }
+
+    /**
+     * Generate a radom number of tasks between 100 and 1000
+     *
+     *
+     * @return the number of tasks created
+     */
+    public int generateRandomTasks() {
+        Random random = new Random();
+        int numberTasks = random.nextInt(901) + 100;
+        for (int i = 0; i < numberTasks; i++) {
+            Task task = createRandomTask();
+            try {
+                taskRepository.saveTask(task);
+            } catch (Exception e) {
+                i--;
+            }
+        }
+        return numberTasks;
+    }
+
+    /**
+     * Create a random task
+     *
+     * @return the task created
+     */
+    private Task createRandomTask() {
+        Random random = new Random();
+
+        String[] levels = {"low", "medium", "high"};
+
+        Task task = new Task();
+
+        task.setNameTask("Random Task" + random.nextInt(1000));
+        task.setDescriptionTask("This is a randomly generated task");
+        task.setDifficultyLevel(levels[random.nextInt(3)]);
+        task.setPriority(random.nextInt(5) + 1);
+        task.setEstimatedTime(LocalDate.now().plusDays(random.nextInt(90) + 1));
+        return task;
+    }
+
+    /**
+     * This method is in charge to delete all existing tasks
+     *
+     * @return the number of tasks deleted
+     */
+    public int deleteAllTasks() {
+        List<Task> tasks = getAllTasks();
+        for (Task task : tasks) {
+            deleteTask(task.getId());
+        }
+        return tasks.size();
+    }
+
+    /**
+     * Method that return how many tasks are in the application
+     *
+     * @return the number of tasks
+     */
+    public int countTasks() {
+        return getAllTasks().size();
     }
 }
